Moved http routes to routes folder, added more data
