+++ backend/index.js
@@ -1,46 +1,13 @@
 import express from "express";
 import { PORT, mongodbURL } from "./config.js";
 import mongoose from "mongoose";
-import { Vinyl } from "./models/vinylModel.js";
+import { vinylsRouter } from "./routes/vinyls.js";
 
 const app = express();
 
 app.use(express.json());
 
-app.post("/vinyls", async (req, res) => {
-	try {
-		if (!req.body.title || !req.body.artist || !req.body.title || !req.body.genres || !req.body.release) {
-			return res.status(400).send({
-				message: "All fields are required (except cover img)",
-			});
-		}
-
-		const newVinyl = {
-			title: req.body.title,
-			artist: req.body.artist,
-			genres: req.body.genres,
-			release: req.body.release,
-			coverImg: req.body.coverImg,
-		};
-
-		const vinyl = await Vinyl.create(newVinyl);
-
-		return res.status(201).send(vinyl);
-	} catch (error) {
-		console.error(error.message);
-		res.status(500).send({ message: error.message });
-	}
-});
-
-app.get("/vinyls", async (req, res) => {
-	try {
-		const vinyls = await Vinyl.find({});
-		return res.status(200).json(vinyls);
-	} catch (error) {
-		console.error(error.message);
-		res.status(500).send({ message: error.message });
-	}
-});
+app.use("/vinyls", vinylsRouter);
 
 mongoose
 	.connect(mongodbURL)
